refactor verifier service
