+++ api/src/services/verifier/sdjwt.ts
@@ -1,0 +1,7 @@
+
+
+export async function verifySDJWT(verifiable: string, nonce?: string, aud?: string): Promise<VerificationResult> {
+
+    return {verified: false};
+
+}
